Don't change cursor on mobile, fix some layout invalidations
This commit disables the custom cursor on devices without a pointer (e.g.
mobile devices). This greatly speeds up especially Chrome on Android (I haven't
tested on iPhone). It uses @media (any-pointer) to check if the pointer is
present, which IE probably does not support. It is possible to use @supports,
but it still does not work in IE. @supports was added in Chrome 28 (2013), where
@media (any-pointer) was added in Chrome 41 (January 2015).

In my mind, undo, redo and other functions without UI prefix should not do any
operations on the DOM, they should only alter the state. Whereas the actual
rendering is done in the UIupdate function as needed.

This commit fixes some layout thrashing. The stones are rendered in the UIupdate
function to not possibly mix DOM reads with DOM writes.

Also, the `move.move` thing was a bit confusing. I'd change "move" in movelist
to something like "entry". I kept "move" for now. It should also be possible in
theory to merge movelist and movedisplay.